Added support for persistence of carpool-parkings in ES but VERY strange library problem .....

+++ src/main/java/be/ictdynamic/mobiscan/services/LoadCarPoolParkingFileService.java
@@ -1,14 +1,22 @@
 package be.ictdynamic.mobiscan.services;
 
 import be.ictdynamic.mobiscan.domain.CarPoolParking;
+import be.ictdynamic.mobiscan.domain.CarPoolParkingES;
 import be.ictdynamic.mobiscan.domain.CarPoolParkingFile;
 import be.ictdynamic.mobiscan.repository.CarPoolParkingRepository;
+import org.elasticsearch.action.index.IndexRequest;
+import org.elasticsearch.action.index.IndexResponse;
+import org.elasticsearch.client.RestHighLevelClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
+import org.json.simple.JSONObject;
+
 import javax.transaction.Transactional;
+import java.io.IOException;
 import java.util.Date;
 
 /**
@@ -22,10 +30,27 @@ public class LoadCarPoolParkingFileService {
     @Autowired
     private CarPoolParkingRepository carPoolParkingRepository;
 
-    public int persistCarPoolParkings(CarPoolParkingFile carPoolParkingFile) {
-        carPoolParkingRepository.deleteAll();
+    @Autowired
+    private RestHighLevelClient restClient;
 
-        carPoolParkingFile.getCarpoolparkings().forEach(carPoolParking -> {
+    @Value("${mobiscan.saveToMySql}")
+    private boolean saveToMySql;
+
+    @Value("${mobiscan.saveToES}")
+    private boolean saveToES;
+
+    @Value("${mobiscan.es.carpool_parking_index_name}")
+    public String ES_CAR_POOL_PARKING_INDEX_NAME;
+
+    @Value("${mobiscan.es.carpool_parking_index_type}")
+    public String ES_CAR_POOL_PARKING_INDEX_TYPE;
+
+    public int persistCarPoolParkings(CarPoolParkingFile carPoolParkingFile) throws IllegalArgumentException, IOException{
+        if (saveToMySql) {
+            carPoolParkingRepository.deleteAll();
+        }
+
+        for (CarPoolParkingFile.CarPoolParking carPoolParking : carPoolParkingFile.getCarpoolparkings()) {
             CarPoolParking carPoolParkingToBePersisted = new CarPoolParking();
             carPoolParkingToBePersisted.setCreatedOn(new Date());
 
@@ -35,10 +60,52 @@ public class LoadCarPoolParkingFileService {
             carPoolParkingToBePersisted.setLatitude(carPoolParking.getParking().getLatitudeOfParking());
             carPoolParkingToBePersisted.setLongitude(carPoolParking.getParking().getLongitudeOfParking());
 
-            carPoolParkingRepository.save(carPoolParkingToBePersisted);
-        });
+            if (saveToMySql) {
+                persistInMySql(carPoolParkingToBePersisted);
+            }
+
+            if (saveToES) {
+                persistInES(carPoolParkingToBePersisted);
+            }
+        };
 
         return carPoolParkingFile.getCarpoolparkings().size();
+    }
+
+    private void persistInES(CarPoolParking carPoolParkingToBePersisted) {
+        carPoolParkingRepository.save(carPoolParkingToBePersisted);
+    }
+
+    private void persistInMySql(CarPoolParking carPoolParkingToBePersisted) throws IllegalArgumentException, IOException {
+        CarPoolParkingES carPoolParkingES = new CarPoolParkingES();
+
+        carPoolParkingES.setCommune(carPoolParkingToBePersisted.getCommune());
+        carPoolParkingES.setType(carPoolParkingToBePersisted.getType());
+        carPoolParkingES.setTitle(carPoolParkingToBePersisted.getTitle());
+        carPoolParkingES.setLatitude(Float.toString(carPoolParkingToBePersisted.getLatitude()));
+        carPoolParkingES.setLongitude(Float.toString(carPoolParkingToBePersisted.getLongitude()));
+
+        JSONObject myLocation = new JSONObject();
+//        myLocation.put("lon", carPoolParkingES.getLongitude());
+//        myLocation.put("lat", carPoolParkingES.getLatitude());
+
+        myLocation.put("lon", "4.4011606");
+        myLocation.put("lat", "51.2061132");
+
+        IndexRequest indexRequest = new IndexRequest(ES_CAR_POOL_PARKING_INDEX_NAME, ES_CAR_POOL_PARKING_INDEX_TYPE)
+                .source(
+                        "createdOn", new Date(),
+                        "commune", carPoolParkingES.getCommune(),
+                        "title", carPoolParkingES.getTitle(),
+                        "type", carPoolParkingES.getType(),
+                        "location", myLocation
+                );
+
+        IndexResponse indexResponse = restClient.index(indexRequest);
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("Persisted esCarPoolParking: {}", carPoolParkingES);
+        }
+
     }
 
 }
